requirements.txt
# Core Dependencies
yfinance
ta
openai
python-dotenv
requests
pydantic-settings
pytest
pytest-mock
tenacity
schedule
python-telegram-bot
feedparser

# API Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Rate Limiting & Performance
slowapi==0.1.9
redis==5.0.1

# Email (AWS SES)
boto3==1.34.34

# HTTP Client (for Tranzila API)
httpx==0.26.0

# Database (Production)
psycopg2-binary==2.9.9

# Monitoring & Error Tracking
sentry-sdk[fastapi]==1.40.0

